+++ logger/logger.go
@@ -15,6 +15,7 @@ var InitLog *logrus.Entry
 var FreecliLog *logrus.Entry
 var SubscriberLog *logrus.Entry
 var GNBLog *logrus.Entry
+var NFLog *logrus.Entry
 
 func init() {
 	log = logrus.New()
@@ -39,6 +40,7 @@ func init() {
 	FreecliLog = log.WithFields(logrus.Fields{"component": "Freecli", "category": "Freecli"})
 	SubscriberLog = log.WithFields(logrus.Fields{"component": "Freecli", "category": "Subscriber Module"})
 	GNBLog = log.WithFields(logrus.Fields{"component": "Freecli", "category": "gNB Module"})
+	NFLog = log.WithFields(logrus.Fields{"component": "Freecli", "category": "NF Module"})
 }
 
 func SetLogLevel(level logrus.Level) {