+++ module/nf/config.go
@@ -1,0 +1,20 @@
+package nf
+
+type Config struct {
+	Info          *Info          `yaml:"info"`
+	Configuration *Configuration `yaml:"configuration"`
+}
+
+type Info struct {
+	Version     string `yaml:"version,omitempty"`
+	Description string `yaml:"description,omitempty"`
+}
+
+type Configuration struct {
+	Mongodb *Mongodb `yaml:"mongodb"`
+}
+
+type Mongodb struct {
+	Name string `yaml:"name"`
+	Url  string `yaml:"url"`
+}