+++ module/nf/const.go
@@ -1,0 +1,4 @@
+package nf
+
+const MODULE_NAME = "nf"
+const MODULE_NF = 0x20