+++ module/nf/nf.go
@@ -1,0 +1,20 @@
+package nf
+
+import (
+	"free5gc-cli/module/nf/api"
+)
+
+var nf *NF
+
+type NF struct {
+}
+
+func (nf *NF) DropDatabase(db string) error {
+	api.Drop(db)
+	return nil
+}
+
+func (nf *NF) FlushDatabase() error {
+	api.Flush()
+	return nil
+}