+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # CHANGELOG
 
+## 2020-11-27
+
+### New
+
+#### Network Function Module
+
+- Database commands
+
+#### QoS Module
+
+- Initialize QoS Module
+
+#### gNB Module
+
+- Register and Deregister user tested against bare-metal and compose version of free5gc v3.0.4
+
+### Improvements
+
+- Clean code of subscriber module
+- Clean code of gNB module
+
 ## 2020-11-26
 
 ### New